refactor(chat): restructure package layout to feature-based MVVM
