+++ app/src/main/java/must/kdroiders/hustlehub/data/local/AppDatabase.kt
@@ -2,11 +2,11 @@ package must.kdroiders.hustlehub.data.local
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import must.kdroiders.hustlehub.data.local.dao.ConversationDao
-import must.kdroiders.hustlehub.data.local.dao.MessageDao
+import must.kdroiders.hustlehub.feature.chat.data.local.dao.ConversationDao
+import must.kdroiders.hustlehub.feature.chat.data.local.dao.MessageDao
 import must.kdroiders.hustlehub.data.local.dao.ServiceDao
-import must.kdroiders.hustlehub.data.local.entity.ConversationEntity
-import must.kdroiders.hustlehub.data.local.entity.MessageEntity
+import must.kdroiders.hustlehub.feature.chat.data.local.entity.ConversationEntity
+import must.kdroiders.hustlehub.feature.chat.data.local.entity.MessageEntity
 import must.kdroiders.hustlehub.data.local.entity.ServiceEntity
 
 @Database(
