+++ app/src/main/java/must/kdroiders/hustlehub/di/AppModule.kt
@@ -12,15 +12,15 @@ import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import must.kdroiders.hustlehub.core.auth.AuthManager
 import must.kdroiders.hustlehub.data.local.AppDatabase
-import must.kdroiders.hustlehub.data.local.dao.ConversationDao
-import must.kdroiders.hustlehub.data.local.dao.MessageDao
+import must.kdroiders.hustlehub.feature.chat.data.local.dao.ConversationDao
+import must.kdroiders.hustlehub.feature.chat.data.local.dao.MessageDao
 import must.kdroiders.hustlehub.data.local.dao.ServiceDao
-import must.kdroiders.hustlehub.data.remote.ChatWebSocketService
-import must.kdroiders.hustlehub.data.remote.ConversationApiService
+import must.kdroiders.hustlehub.feature.chat.data.remote.ChatWebSocketService
+import must.kdroiders.hustlehub.feature.chat.data.remote.ConversationApiService
 import must.kdroiders.hustlehub.data.remote.MediaApiService
 import must.kdroiders.hustlehub.data.remote.ServiceApiService
 import must.kdroiders.hustlehub.data.remote.UserApiService
-import must.kdroiders.hustlehub.data.repository.ChatRepositoryImpl
+import must.kdroiders.hustlehub.feature.chat.data.repository.ChatRepositoryImpl
 import must.kdroiders.hustlehub.data.repository.ServiceRepositoryImpl
 import must.kdroiders.hustlehub.data.repository.StorageRepository
 import must.kdroiders.hustlehub.data.repository.StorageRepositoryImpl
@@ -28,7 +28,7 @@ import must.kdroiders.hustlehub.data.repository.UserRepository
 import must.kdroiders.hustlehub.data.repository.UserRepositoryImpl
 import must.kdroiders.hustlehub.datastore.UserPreferences
 import must.kdroiders.hustlehub.datastore.dataStore
-import must.kdroiders.hustlehub.domain.repository.ChatRepository
+import must.kdroiders.hustlehub.feature.chat.domain.repository.ChatRepository
 import must.kdroiders.hustlehub.domain.repository.ServiceRepository
 import must.kdroiders.hustlehub.ui.features.auth.data.remote.AuthApiService
 import must.kdroiders.hustlehub.ui.features.auth.data.repository.AuthRepositoryImpl
