+++ app/src/main/java/must/kdroiders/hustlehub/di/NetworkModule.kt
@@ -9,7 +9,7 @@ import must.kdroiders.hustlehub.BuildConfig
 import must.kdroiders.hustlehub.core.api.AuthInterceptor
 import must.kdroiders.hustlehub.core.api.TokenAuthenticator
 import must.kdroiders.hustlehub.core.auth.AuthManager
-import must.kdroiders.hustlehub.data.remote.ConversationApiService
+import must.kdroiders.hustlehub.feature.chat.data.remote.ConversationApiService
 import must.kdroiders.hustlehub.data.remote.MediaApiService
 import must.kdroiders.hustlehub.data.remote.ServiceApiService
 import must.kdroiders.hustlehub.data.remote.UserApiService
