+++ app/src/main/java/must/kdroiders/hustlehub/feature/chat/domain/model/Conversation.kt
@@ -1,4 +1,4 @@
-package must.kdroiders.hustlehub.data.model
+package must.kdroiders.hustlehub.feature.chat.domain.model
 
 data class Conversation(
     val id: String,
