+++ app/src/main/java/must/kdroiders/hustlehub/feature/chat/presentation/audio/VoiceRecorder.kt
@@ -1,4 +1,4 @@
-package must.kdroiders.hustlehub.ui.features.chat.audio
+package must.kdroiders.hustlehub.feature.chat.presentation.audio
 
 import android.content.Context
 import android.media.MediaRecorder
