+++ app/src/main/java/must/kdroiders/hustlehub/navigation/HustleHubNavGraph.kt
@@ -37,7 +37,7 @@ import must.kdroiders.hustlehub.ui.features.auth.presentation.viewmodel.LoginVie
 import must.kdroiders.hustlehub.ui.features.profilesetup.presentation.view.ProfileSetupScreen
 import must.kdroiders.hustlehub.ui.features.service.presentation.view.CreateServiceScreen
 import must.kdroiders.hustlehub.ui.features.service.presentation.view.MyServicesScreen
-import must.kdroiders.hustlehub.ui.features.chat.ChatDetailScreen
+import must.kdroiders.hustlehub.feature.chat.presentation.view.ChatDetailScreen
 import must.kdroiders.hustlehub.ui.features.settings.presentation.view.SettingsScreen
 
 /**
